Moved new test to parsing.tests.
